backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
langchain==1.2.15
langchain-community==0.4.1
langchain-google-genai==4.2.2
faiss-cpu==1.13.2
python-dotenv==1.2.2
pydantic-settings==2.13.1
google-auth==2.49.2
